update: Refactored MealPlanScreen to be modular
Next step is look at type safety and how we can better optimize the
perfomance
